Rename update_common module to common
