+++ app/shell/py/pie/pie/update/pubdate.py
@@ -5,7 +5,7 @@ from pathlib import Path
 from typing import Iterable, Sequence
 
 from pie.logging import add_log_argument, setup_file_logger
-from pie.update_common import get_changed_files, update_files as update_common_files
+from .common import get_changed_files, update_files as common_update_files
 from pie.utils import get_pubdate
 
 
@@ -28,7 +28,7 @@ def update_files(paths: Iterable[Path], pubdate: str) -> tuple[list[str], int]:
     entries for each modified file and ``checked`` is the number of files that
     were examined.
     """
-    return update_common_files(paths, "pubdate", pubdate)
+    return common_update_files(paths, "pubdate", pubdate)
 
 
 def main(argv: Sequence[str] | None = None) -> int:
